pyproject.toml
[tool.poetry]
name = "own-game"
version = "0.1.1"
description = ""
authors = ["Kirill Lapushinskiy <kirill@lapushinskiy.ru>"]
readme = "README.md"
packages = [{ include = "own_game" }]


[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.4"
aiohttp-apispec = "^2.2.3"
aiohttp-session = { extras = ["secure"], version = "^2.12.0" }
sqlalchemy = "^2.0.4"
pyyaml = "^6.0"
cryptography = "^39.0.1"
orjson = "^3.8.6"
dacite = "^1.8.0"
asyncpg = "^0.27.0"
aiolimiter = "^1.0.0"
alembic = "^1.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-asyncio = "^0.20.3"
aiosqlite = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
